requirements.txt
requests==2.23.0
liquidpy==0.5.0